+++ axes/middleware.py
@@ -1,12 +1,25 @@
 from typing import Callable
 
-from axes.helpers import get_lockout_response
+from django.conf import settings
+
+from axes.helpers import (
+    get_lockout_response,
+    get_failure_limit,
+    get_client_username,
+    get_credentials,
+)
+
+from axes.handlers.proxy import AxesProxyHandler
 
 
 class AxesMiddleware:
     """
     Middleware that calculates necessary HTTP request attributes for attempt monitoring
     and maps lockout signals into readable HTTP 403 Forbidden responses.
+
+    If a project uses ``django rest framework`` then the middleware updates the
+    request and checks whether the limit has been exceeded. It's needed only
+    for integration with DRF because it uses its own request object.
 
     This middleware recognizes a logout monitoring flag in the request and
     and uses the ``axes.helpers.get_lockout_response`` handler for returning
@@ -29,7 +42,22 @@ class AxesMiddleware:
     def __call__(self, request):
         response = self.get_response(request)
 
-        if getattr(request, "axes_locked_out", None):
-            response = get_lockout_response(request)  # type: ignore
+        if settings.AXES_ENABLED:
+            if "rest_framework" in settings.INSTALLED_APPS:
+                AxesProxyHandler.update_request(request)
+                username = get_client_username(request)
+                credentials = get_credentials(username)
+                failures_since_start = AxesProxyHandler.get_failures(
+                    request, credentials
+                )
+                if (
+                    settings.AXES_LOCK_OUT_AT_FAILURE
+                    and failures_since_start >= get_failure_limit(request, credentials)
+                ):
+
+                    request.axes_locked_out = True
+
+            if getattr(request, "axes_locked_out", None):
+                response = get_lockout_response(request)  # type: ignore
 
         return response