+++ CHANGES.rst
@@ -3,6 +3,33 @@ Changes
 =======
 
 
+5.9.1 (2020-12-02)
+------------------
+
+- Move tests to GitHub Actions
+  [jezdez]
+- Fix running Axes code in middleware when ``AXES_ENABLED`` is ``False``.
+  [ashokdelphia]
+
+
+5.9.0 (2020-11-05)
+------------------
+
+- Add Python 3.9 support.
+  [hramezani]
+- Prevent ``AccessAttempt`` creation with database handler 
+  when username is not set
+  and ``AXES_ONLY_USER_FAILURES`` setting is not set.
+  [hramezani]
+
+
+5.8.0 (2020-10-16)
+------------------
+
+- Improve Django REST Framework (DRF) integration.
+  [Anatoly]
+
+
 5.7.1 (2020-09-27)
 ------------------